requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
discord
docutils==0.16
entrypoints==0.3
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.1
pickle-mixin
pickleshare==0.7.5
Pillow==7.1.2
playsound==1.2.2
prometheus-client==0.9.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
pynacl
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pyspellchecker
pytz==2020.1
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
rstcheck==3.3.1
Send2Trash==1.5.0
six==1.15.0
sounddevice==0.3.15
soupsieve==2.0
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0